requirements.txt
rdflib>=7.0
python-dateutil>=2.8.2
pyoxigraph
kafka-python
requests
pytz
